# Refactor and Optimization of Match Resource Definition